add some fixes

+++ scripts/zsh/zsh.sh
@@ -1,3 +1,10 @@
+#!/bin/bash
+
+install_zplug () {
+    curl -sL --proto-redir -all,https https://raw.githubusercontent.com/zplug/installer/master/installer.zsh | zsh
+}
+
+sleep 1
 printf "\n\n"
 printf "_______________________________________________________________________________"
 printf "\n\nYou are going to install zsh with some plugins.\nYou have 10 seconds to cancel this action\n"
@@ -9,25 +16,22 @@ apt update
 printf "ZSH and oh-my-zsh\n"
 apt install -y curl zsh git
 curl https://raw.githubusercontent.com/padawarmik/docker-scripts/main/scripts/zsh/.p10k.zsh -o ~/.p10k.zsh -s
-curl https://raw.githubusercontent.com/padawarmik/docker-scripts/zplug/scripts/zsh/.zshrc -o ~/.zshrc -s
+if [ -f ~/.zshrc ]
+then
+    read -p "Your .zshrc file will be replaced. Do you want to continue?" -n 1 -r
+    echo
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        rm ~/.zshrc -f
+        curl https://raw.githubusercontent.com/padawarmik/docker-scripts/zplug/scripts/zsh/.zshrc -o ~/.zshrc -s
+    fi
+fi
 sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 
-curl -sL --proto-redir -all,https https://raw.githubusercontent.com/zplug/installer/master/installer.zsh | zsh
-
-# printf "Now I am going to clone requested plugins\n"
-# sleep 1
-# git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-# git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
-# git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
-# printf "Changing of ~/.zshrc to apply changes\n"
-# sed -i "/ZSH_THEME=/c\ZSH_THEME=\"powerlevel10k/powerlevel10k\"" ~/.zshrc
-# sed -i "/plugins=/c\plugins=(git zsh-autosuggestions zsh-syntax-highlighting)" ~/.zshrc
-# grep -q ZSH_AUTOSUGGEST_STRATEGY ~/.zshrc
-# if [ $? -eq 1 ]
-# then
-#     echo "ZSH_AUTOSUGGEST_STRATEGY=" >> ~/.zshrc
-# fi
-# sed -i "/ZSH_AUTOSUGGEST_STRATEGY=/c\ZSH_AUTOSUGGEST_STRATEGY=(history completion)" ~/.zshrc
+if [ ! -d "~/.zplug" ]
+then
+    install_zplug()
+fi
 chsh -s $(which zsh)
 if [ $SHELL == "/usr/bin/zsh" ]
 then
